Add Custom hook of Registration page validation
