+++ README.md
@@ -138,10 +138,7 @@
 
 #### [ 코드리뷰 반영(스프린트 미션 4) ]
 
-- [ ] baseURL 변수명 변경
-- [ ] api, params 동작 검증 로직 추가
-- [ ] .catch 에러 메시지에 응답 코드 추가(별도 함수 관리)
-- [ ] get api를 하나로 병합
+- [x] baseURL 변수명 변경
 
 #### [ 추가 ]
 
@@ -173,10 +170,8 @@
 
 #### [ 코드리뷰 반영(스프린트 미션 5) ]
 
-- [ ]
-- [ ]
-- [ ]
-- [ ]
+- [ ] pagination을 context로 표현하기
+- [ ] select를 custom hook으로 만들기
 
 #### [ 추가 ]
 
@@ -184,6 +179,7 @@
 - 데이터베이스 생성 & 연결
 - 상품 등록 페이지
 - API 연결
+- 상품 등록 유효성 검사 커스텀 훅
 
 #### [ 수정 ]
